finished restructuring unit tests
