--- batchglm/unit_test/glm_all/external.py
@@ -1,4 +1,0 @@
-from batchglm.unit_test.base_glm import Test_AccuracySizeFactors_GLM, _Test_AccuracySizeFactors_GLM_Estim
-from batchglm.unit_test.base_glm import Test_Graph_GLM, _TestGraphGLMEstim
-from batchglm.unit_test.base_glm import Test_DataTypes_GLM, _Test_DataTypes_GLM_Estim
-from batchglm.unit_test.base_glm import Test_ExtremValues_GLM, _Test_ExtremValues_GLM_Estim
