+++ batchglm/unit_test/test_acc_constrained_vglm_all.py
@@ -3,28 +3,43 @@ import numpy as np
 import unittest
 
 import batchglm.api as glm
-from batchglm.unit_test.glm_all.test_acc_glm_all import _TestAccuracyGlmAll
+from batchglm.unit_test.test_acc_glm_all import _TestAccuracyGlmAll
 
 glm.setup_logging(verbosity="WARNING", stream="STDOUT")
 logger = logging.getLogger(__name__)
 
 
-class _TestAccuracyGlmAllSf(_TestAccuracyGlmAll):
+class _TestAccuracyVglmAll(_TestAccuracyGlmAll):
 
     def simulate(self):
         super().simulate()
-        # Add size factors into input data: Do not centre at 1 so that they bias MAD if something is off.
-        self.sim1.input_data.size_factors = np.random.uniform(1.5, 2., size=self.sim1.input_data.num_observations)
+        # Override design matrix of simulation 1 to encode constraints
+        dmat = np.hstack([
+            self.sim1.input_data.design_loc,
+            np.expand_dims(self.sim1.input_data.design_loc[:, 0] -
+                           self.sim1.input_data.design_loc[:, -1], axis=-1)
+        ])
+        constraints = np.zeros([4, 3])
+        constraints[0, 0] = 1
+        constraints[1, 1] = 1
+        constraints[2, 2] = 1
+        constraints[3, 2] = -1
+        self.sim1.input_data.design_loc = dmat
+        self.sim1.input_data.design_scale = dmat
+        self.sim1.input_data.constraints_loc = constraints
+        self.sim1.input_data.constraints_scale = constraints
 
     def _test_full(self, sparse):
         self._test_full_a_and_b(sparse=sparse)
+        self._test_full_a_only(sparse=sparse)
 
     def _test_batched(self, sparse):
         self._test_batched_a_and_b(sparse=sparse)
+        self._test_batched_a_only(sparse=sparse)
 
 
-class TestAccuracyGlmNbSf(
-    _TestAccuracyGlmAllSf,
+class TestAccuracyVglmNb(
+    _TestAccuracyVglmAll,
     unittest.TestCase
 ):
     """
@@ -34,7 +49,7 @@ class TestAccuracyGlmNbSf(
     def test_full_nb(self):
         logging.getLogger("tensorflow").setLevel(logging.INFO)
         logging.getLogger("batchglm").setLevel(logging.INFO)
-        logger.error("TestAccuracyGlmNbSf.test_full_nb()")
+        logger.error("TestAccuracyVglmNb.test_full_nb()")
 
         np.random.seed(1)
         self.noise_model = "nb"
@@ -45,7 +60,7 @@ class TestAccuracyGlmNbSf(
     def test_batched_nb(self):
         logging.getLogger("tensorflow").setLevel(logging.INFO)
         logging.getLogger("batchglm").setLevel(logging.INFO)
-        logger.error("TestAccuracyGlmNbSf.test_batched_nb()")
+        logger.error("TestAccuracyVglmNb.test_batched_nb()")
 
         np.random.seed(1)
         self.noise_model = "nb"
@@ -54,7 +69,7 @@ class TestAccuracyGlmNbSf(
         self._test_batched(sparse=True)
 
 
-class TestAccuracyGlmNormSf(
+class TestAccuracyVglmNorm(
     _TestAccuracyGlmAll,
     unittest.TestCase
 ):
@@ -66,7 +81,7 @@ class TestAccuracyGlmNormSf(
     def test_full_norm(self):
         logging.getLogger("tensorflow").setLevel(logging.INFO)
         logging.getLogger("batchglm").setLevel(logging.INFO)
-        logger.error("TestAccuracyGlmNormSf.test_full_norm()")
+        logger.error("TestAccuracyVglmNorm.test_full_norm()")
 
         np.random.seed(1)
         self.noise_model = "norm"
@@ -77,7 +92,7 @@ class TestAccuracyGlmNormSf(
     def test_batched_norm(self):
         logging.getLogger("tensorflow").setLevel(logging.INFO)
         logging.getLogger("batchglm").setLevel(logging.INFO)
-        logger.error("TestAccuracyGlmNormSf.test_batched_norm()")
+        logger.error("TestAccuracyVglmNorm.test_batched_norm()")
 
         np.random.seed(1)
         self.noise_model = "norm"
@@ -86,17 +101,36 @@ class TestAccuracyGlmNormSf(
         self._test_batched(sparse=True)
 
 
-class TestAccuracyGlmBetaSf(
+class TestAccuracyVglmBeta(
     _TestAccuracyGlmAll,
     unittest.TestCase
 ):
     """
     Test whether optimizers yield exact results for beta distributed data.
-    Note: size factors are note implemented for beta distribution.
+    TODO not working yet.
     """
 
-    def test_dummy(self):
-        return True
+    def test_full_beta(self):
+        logging.getLogger("tensorflow").setLevel(logging.INFO)
+        logging.getLogger("batchglm").setLevel(logging.INFO)
+        logger.error("TestAccuracyVglmBeta.test_full_beta()")
+
+        np.random.seed(1)
+        self.noise_model = "beta"
+        self.simulate()
+        self._test_full(sparse=False)
+        self._test_full(sparse=True)
+
+    def test_batched_beta(self):
+        logging.getLogger("tensorflow").setLevel(logging.INFO)
+        logging.getLogger("batchglm").setLevel(logging.INFO)
+        logger.error("TestAccuracyVglmBeta.test_batched_beta()")
+
+        np.random.seed(1)
+        self.noise_model = "beta"
+        self.simulate()
+        self._test_batched(sparse=False)
+        self._test_batched(sparse=True)
 
 
 if __name__ == '__main__':
