+++ batchglm/unit_test/test_acc_glm_all.py
