+++ batchglm/unit_test/test_acc_sizefactors_glm_all.py
@@ -3,43 +3,28 @@ import numpy as np
 import unittest
 
 import batchglm.api as glm
-from batchglm.unit_test.glm_all.test_acc_glm_all import _TestAccuracyGlmAll
+from batchglm.unit_test.test_acc_glm_all import _TestAccuracyGlmAll
 
 glm.setup_logging(verbosity="WARNING", stream="STDOUT")
 logger = logging.getLogger(__name__)
 
 
-class _TestAccuracyVglmAll(_TestAccuracyGlmAll):
+class _TestAccuracyGlmAllSf(_TestAccuracyGlmAll):
 
     def simulate(self):
         super().simulate()
-        # Override design matrix of simulation 1 to encode constraints
-        dmat = np.hstack([
-            self.sim1.input_data.design_loc,
-            np.expand_dims(self.sim1.input_data.design_loc[:, 0] -
-                           self.sim1.input_data.design_loc[:, -1], axis=-1)
-        ])
-        constraints = np.zeros([4, 3])
-        constraints[0, 0] = 1
-        constraints[1, 1] = 1
-        constraints[2, 2] = 1
-        constraints[3, 2] = -1
-        self.sim1.input_data.design_loc = dmat
-        self.sim1.input_data.design_scale = dmat
-        self.sim1.input_data.constraints_loc = constraints
-        self.sim1.input_data.constraints_scale = constraints
+        # Add size factors into input data: Do not centre at 1 so that they bias MAD if something is off.
+        self.sim1.input_data.size_factors = np.random.uniform(1.5, 2., size=self.sim1.input_data.num_observations)
 
     def _test_full(self, sparse):
         self._test_full_a_and_b(sparse=sparse)
-        self._test_full_a_only(sparse=sparse)
 
     def _test_batched(self, sparse):
         self._test_batched_a_and_b(sparse=sparse)
-        self._test_batched_a_only(sparse=sparse)
 
 
-class TestAccuracyVglmNb(
-    _TestAccuracyVglmAll,
+class TestAccuracyGlmNbSf(
+    _TestAccuracyGlmAllSf,
     unittest.TestCase
 ):
     """
@@ -49,7 +34,7 @@ class TestAccuracyVglmNb(
     def test_full_nb(self):
         logging.getLogger("tensorflow").setLevel(logging.INFO)
         logging.getLogger("batchglm").setLevel(logging.INFO)
-        logger.error("TestAccuracyVglmNb.test_full_nb()")
+        logger.error("TestAccuracyGlmNbSf.test_full_nb()")
 
         np.random.seed(1)
         self.noise_model = "nb"
@@ -60,7 +45,7 @@ class TestAccuracyVglmNb(
     def test_batched_nb(self):
         logging.getLogger("tensorflow").setLevel(logging.INFO)
         logging.getLogger("batchglm").setLevel(logging.INFO)
-        logger.error("TestAccuracyVglmNb.test_batched_nb()")
+        logger.error("TestAccuracyGlmNbSf.test_batched_nb()")
 
         np.random.seed(1)
         self.noise_model = "nb"
@@ -69,7 +54,7 @@ class TestAccuracyVglmNb(
         self._test_batched(sparse=True)
 
 
-class TestAccuracyVglmNorm(
+class TestAccuracyGlmNormSf(
     _TestAccuracyGlmAll,
     unittest.TestCase
 ):
@@ -81,7 +66,7 @@ class TestAccuracyVglmNorm(
     def test_full_norm(self):
         logging.getLogger("tensorflow").setLevel(logging.INFO)
         logging.getLogger("batchglm").setLevel(logging.INFO)
-        logger.error("TestAccuracyVglmNorm.test_full_norm()")
+        logger.error("TestAccuracyGlmNormSf.test_full_norm()")
 
         np.random.seed(1)
         self.noise_model = "norm"
@@ -92,7 +77,7 @@ class TestAccuracyVglmNorm(
     def test_batched_norm(self):
         logging.getLogger("tensorflow").setLevel(logging.INFO)
         logging.getLogger("batchglm").setLevel(logging.INFO)
-        logger.error("TestAccuracyVglmNorm.test_batched_norm()")
+        logger.error("TestAccuracyGlmNormSf.test_batched_norm()")
 
         np.random.seed(1)
         self.noise_model = "norm"
@@ -101,36 +86,17 @@ class TestAccuracyVglmNorm(
         self._test_batched(sparse=True)
 
 
-class TestAccuracyVglmBeta(
+class TestAccuracyGlmBetaSf(
     _TestAccuracyGlmAll,
     unittest.TestCase
 ):
     """
     Test whether optimizers yield exact results for beta distributed data.
-    TODO not working yet.
+    Note: size factors are note implemented for beta distribution.
     """
 
-    def test_full_beta(self):
-        logging.getLogger("tensorflow").setLevel(logging.INFO)
-        logging.getLogger("batchglm").setLevel(logging.INFO)
-        logger.error("TestAccuracyVglmBeta.test_full_beta()")
-
-        np.random.seed(1)
-        self.noise_model = "beta"
-        self.simulate()
-        self._test_full(sparse=False)
-        self._test_full(sparse=True)
-
-    def test_batched_beta(self):
-        logging.getLogger("tensorflow").setLevel(logging.INFO)
-        logging.getLogger("batchglm").setLevel(logging.INFO)
-        logger.error("TestAccuracyVglmBeta.test_batched_beta()")
-
-        np.random.seed(1)
-        self.noise_model = "beta"
-        self.simulate()
-        self._test_batched(sparse=False)
-        self._test_batched(sparse=True)
+    def test_dummy(self):
+        return True
 
 
 if __name__ == '__main__':
