+++ batchglm/unit_test/test_hessians_glm_all.py
